fill some sorries and suggest lemma (#18)
filled sorries of PrimeSpectrum.connectedSpace_of_faithfullyFlat and
PrimeSpectrum.connectedSpace_of_isScalarTower
As the latter works just like
PrimeSpectrum.irreducibleSpace_of_isScalarTower (which was done by
@TimoKraenzle), I added a lemma baseChange_along_field_extensions (name
to be improved) in Algebra Irreducible abstracting what was done in that
proof. I don't know what the best way is to set up this lemma (I had to
declare some local instances). If this is sensible, we could shorten the
proof of PrimeSpectrum.irreducibleSpace_of_isScalarTower.